🔄 refactor: restructure site detail pages with layout pattern
- Transform single page.tsx into layout.tsx with tabs navigation
- Create individual pages for batiments, etages, parties, lots
- Add SiteProvider for shared state management
- Follow same pattern as administration section
- Maintain all existing functionality with cleaner architecture
