+++ src/app/dashboard/(registre-securite)/sites/[siteId]/lots/page.tsx
@@ -1,0 +1,19 @@
+'use client';
+
+import { useParams } from 'next/navigation';
+import LotsTab from '../_components/LotsTab';
+import { useSiteContext } from '../_providers/SiteProvider';
+
+export default function LotsPage() {
+  const params = useParams();
+  const siteId = parseInt(params.siteId as string);
+  const { isDeleted, showNotification } = useSiteContext();
+
+  return (
+    <LotsTab
+      siteId={siteId}
+      onNotification={showNotification}
+      disabled={isDeleted}
+    />
+  );
+}
